Pr 147/rename item categories to departments

+++ database/migrations/2016_07_09_114515_create_item_category_table.php
@@ -28,6 +28,8 @@ class CreateItemCategoryTable extends Migration
      */
     public function down()
     {
-        Schema::drop('item_categories');
+        if(Schema::hasTable('item_categories')) {
+            Schema::drop('item_categories');
+        }
     }
 }
